\name{as.gt}
\alias{as.gt}
\title{Extract genotype profile for a single contributor from rGTs output}
\description{Extract genotype profile for a single contributor from rGTs output}
\usage{
as.gt(res,ind)
}
\arguments{
  \item{res}{Output from rGTs}
  \item{ind}{Integer, which individual's genotype profile should be extracted}
}
\value{Data frame, genotype profile for selected individual, for format see \code{\link{formats}}.}
\author{
Peter Green (P.J.Green@bristol.ac.uk)
}

