\name{emperors}
\alias{emperors}
\docType{data}
\title{IBD pattern distribution in the Iulius-Claudius pedigree}
\description{IBD pattern distribution in the Iulius-Claudius pedigree}
\usage{data("emperors")}
\format{
IBD object
}
\seealso{\code{\link{formats}}.}
\examples{
data(emperors)
}
\keyword{datasets}
