\name{intoMix}
\alias{intoMix}
\title{Edit output from rGTs to omit individuals with NA amounts of DNA}
\description{Edit output from rGTs to omit individuals with NA amounts of DNA}
\usage{
intoMix(res)
}
\arguments{
  \item{res}{Output from \code{rGTs}}
}
\value{The edited data structure}
\author{
Peter Green (P.J.Green@bristol.ac.uk)
}
