\name{test2data}
\alias{test2data}
\alias{Cgt} 
\alias{Fgt} 
\alias{Mgt} 
\alias{Rgt} 
\alias{S1gt} 
\alias{S2gt} 
\alias{db} 
\alias{epg}
\docType{data}
\title{Small data set for demonstrating some capabilities of KinMix and KinMixLite}
\description{Small test data set (2 markers with 4 or 5 alleles each, plus AMEL), for demonstrating some capabilities of KinMix  and KinMixLite}
\usage{data("test2data")}
\format{
Data objects for demonstrating KinMix:
\code{epg} (DNAmixturesLite peak height data), 
\code{db} (DNAmixturesLite allele frequency database), 
and \code{Cgt, Fgt, Mgt, Rgt, S1gt, S2gt} potential relative genotype data frames.
}
\examples{
data(test2data)
}
\keyword{datasets}
